# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Quinn's Reading List API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module. <a href='https://github.com/qcoyle/books'>GitHub</a>"
  },
  "host": "api.quinncoyle.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Book",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "See all books",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Book"
        ],
        "description": "Add a book",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "Get book by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Book"
        ],
        "description": "Modify a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "description": "Delete a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "definitions": {}
}